+++ plugin/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.23](https://github.com/nativescript-community/ui-label/compare/v1.2.22...v1.2.23) (2023-02-25)
+
+**Note:** Version bump only for package @nativescript-community/ui-label
+
+## [1.2.22](https://github.com/nativescript-community/ui-label/compare/v1.2.21...v1.2.22) (2023-02-20)
+
+**Note:** Version bump only for package @nativescript-community/ui-label
+
 ## [1.2.21](https://github.com/nativescript-community/ui-label/compare/v1.2.20...v1.2.21) (2023-01-30)
 
 **Note:** Version bump only for package @nativescript-community/ui-label